HKSampleGraph+ViewModel
